feat: modify resolver from create/update of local authority profiles
